Uncomment some client classes, but still in progress. -CommandUtilTest now have SellerCommandTestUtil and BuyerCommandTestUtil -Will try to delete Typical Client later, but the storage part hasn't been changed yet -EditClientDescriptorBuilder will try to extract into EditSellerDescriptorBuilder and EditBuyerDescriptorBuilder (not yet done)
