--- src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -1,207 +1,0 @@
-package seedu.address.logic.commands;
-
-import static java.util.Objects.requireNonNull;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
-import static seedu.address.model.Model.PREDICATE_SHOW_ALL_CLIENTS;
-
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-
-import seedu.address.commons.core.Messages;
-import seedu.address.commons.core.index.Index;
-import seedu.address.commons.util.CollectionUtil;
-import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
-import seedu.address.model.client.Appointment;
-import seedu.address.model.client.Client;
-import seedu.address.model.client.Name;
-import seedu.address.model.client.Phone;
-import seedu.address.model.tag.Tag;
-
-/**
- * Edits the details of an existing client in the address book.
- */
-public class EditCommand extends Command {
-
-    public static final String COMMAND_WORD = "edit";
-
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Edits the details of the client identified "
-            + "by the index number used in the displayed client list. "
-            + "Existing values will be overwritten by the input values.\n"
-            + "Parameters: INDEX (must be a positive integer) "
-            + "[" + PREFIX_NAME + "NAME] "
-            + "[" + PREFIX_PHONE + "PHONE] "
-            + "[" + PREFIX_TAG + "TAG]...\n"
-            + "Example: " + COMMAND_WORD + " 1 "
-            + PREFIX_PHONE + "91234567 ";
-
-    public static final String MESSAGE_EDIT_CLIENT_SUCCESS = "Edited client: %1$s";
-    public static final String MESSAGE_NOT_EDITED = "At least one field to edit must be provided.";
-    public static final String MESSAGE_DUPLICATE_CLIENTS = "This client already exists in the address book.";
-
-    private final Index index;
-    private final EditclientDescriptor editclientDescriptor;
-
-    /**
-     * @param index of the client in the filtered client list to edit
-     * @param editclientDescriptor details to edit the client with
-     */
-    public EditCommand(Index index, EditclientDescriptor editclientDescriptor) {
-        requireNonNull(index);
-        requireNonNull(editclientDescriptor);
-
-        this.index = index;
-        this.editclientDescriptor = new EditclientDescriptor(editclientDescriptor);
-    }
-
-    @Override
-    public CommandResult execute(Model model) throws CommandException {
-        requireNonNull(model);
-        List<Client> lastShownList = model.getFilteredClientList();
-
-        if (index.getZeroBased() >= lastShownList.size()) {
-            throw new CommandException(Messages.MESSAGE_INVALID_CLIENT_DISPLAYED_INDEX);
-        }
-
-        Client clientToEdit = lastShownList.get(index.getZeroBased());
-        Client editedClient = createEditedclient(clientToEdit, editclientDescriptor);
-
-        if (!clientToEdit.isSameclient(editedClient) && model.hasClient(editedClient)) {
-            throw new CommandException(MESSAGE_DUPLICATE_CLIENTS);
-        }
-
-        model.setClient(clientToEdit, editedClient);
-        model.updateFilteredClientList(PREDICATE_SHOW_ALL_CLIENTS);
-        return new CommandResult(String.format(MESSAGE_EDIT_CLIENT_SUCCESS, editedClient));
-    }
-
-    /**
-     * Creates and returns a {@code client} with the details of {@code clientToEdit}
-     * edited with {@code editclientDescriptor}.
-     */
-    private static Client createEditedclient(Client clientToEdit, EditclientDescriptor editclientDescriptor) {
-        assert clientToEdit != null;
-
-        Name updatedName = editclientDescriptor.getName().orElse(clientToEdit.getName());
-        Phone updatedPhone = editclientDescriptor.getPhone().orElse(clientToEdit.getPhone());
-        Set<Tag> updatedTags = editclientDescriptor.getTags().orElse(clientToEdit.getTags());
-        Appointment updatedAppointment = editclientDescriptor.getAppointment().orElse(clientToEdit.getAppointment());
-
-        return new Client(updatedName, updatedPhone, updatedAppointment, updatedTags);
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        // short circuit if same object
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof EditCommand)) {
-            return false;
-        }
-
-        // state check
-        EditCommand e = (EditCommand) other;
-        return index.equals(e.index)
-                && editclientDescriptor.equals(e.editclientDescriptor);
-    }
-
-    /**
-     * Stores the details to edit the client with. Each non-empty field value will replace the
-     * corresponding field value of the client.
-     */
-    public static class EditclientDescriptor {
-        private Name name;
-        private Phone phone;
-        private Set<Tag> tags;
-        private Appointment appointment;
-
-        public EditclientDescriptor() {}
-
-        /**
-         * Copy constructor.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public EditclientDescriptor(EditclientDescriptor toCopy) {
-            setName(toCopy.name);
-            setPhone(toCopy.phone);
-            setTags(toCopy.tags);
-        }
-
-
-        /**
-         * Returns true if at least one field is edited.
-         */
-        public boolean isAnyFieldEdited() {
-            return CollectionUtil.isAnyNonNull(name, phone, tags);
-        }
-
-        public void setName(Name name) {
-            this.name = name;
-        }
-
-        public Optional<Name> getName() {
-            return Optional.ofNullable(name);
-        }
-
-        public void setPhone(Phone phone) {
-            this.phone = phone;
-        }
-
-        public Optional<Phone> getPhone() {
-            return Optional.ofNullable(phone);
-        }
-
-        public void setAppointment(Appointment appointment) {
-            this.appointment = appointment;
-        }
-
-        public Optional<Appointment> getAppointment() {
-            return Optional.ofNullable(appointment);
-        }
-
-        /**
-         * Sets {@code tags} to this object's {@code tags}.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public void setTags(Set<Tag> tags) {
-            this.tags = (tags != null) ? new HashSet<>(tags) : null;
-        }
-
-        /**
-         * Returns an unmodifiable tag set, which throws {@code UnsupportedOperationException}
-         * if modification is attempted.
-         * Returns {@code Optional#empty()} if {@code tags} is null.
-         */
-        public Optional<Set<Tag>> getTags() {
-            return (tags != null) ? Optional.of(Collections.unmodifiableSet(tags)) : Optional.empty();
-        }
-
-        @Override
-        public boolean equals(Object other) {
-            // short circuit if same object
-            if (other == this) {
-                return true;
-            }
-
-            // instanceof handles nulls
-            if (!(other instanceof EditclientDescriptor)) {
-                return false;
-            }
-
-            // state check
-            EditclientDescriptor e = (EditclientDescriptor) other;
-
-            return getName().equals(e.getName())
-                    && getPhone().equals(e.getPhone())
-                    && getTags().equals(e.getTags());
-        }
-    }
-}
