--- src/test/java/seedu/address/logic/commands/AddCommandIntegrationTest.java
@@ -1,49 +1,0 @@
-package seedu.address.logic.commands;
-
-import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
-import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.testutil.TypicalClients.getTypicalAddressBook;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import seedu.address.model.BuyerAddressBook;
-import seedu.address.model.Model;
-import seedu.address.model.ModelManager;
-import seedu.address.model.SellerAddressBook;
-import seedu.address.model.UserPrefs;
-import seedu.address.model.client.Client;
-import seedu.address.testutil.ClientBuilder;
-
-/**
- * Contains integration tests (interaction with the Model) for {@code AddCommand}.
- */
-public class AddCommandIntegrationTest {
-
-    private Model model;
-
-    @BeforeEach
-    public void setUp() {
-        model = new ModelManager(getTypicalAddressBook(), new UserPrefs(), new SellerAddressBook(),
-                new BuyerAddressBook());
-    }
-
-    @Test
-    public void execute_newclient_success() {
-        Client validClient = new ClientBuilder().build();
-
-        Model expectedModel = new ModelManager(model.getAddressBook(), new UserPrefs(), new SellerAddressBook(),
-                new BuyerAddressBook());
-        expectedModel.addClient(validClient);
-
-        assertCommandSuccess(new AddCommand(validClient), model,
-                String.format(AddCommand.MESSAGE_SUCCESS, validClient), expectedModel);
-    }
-
-    @Test
-    public void execute_duplicateclient_throwsCommandException() {
-        Client clientInList = model.getAddressBook().getclientList().get(0);
-        assertCommandFailure(new AddCommand(clientInList), model, AddCommand.MESSAGE_DUPLICATE_CLIENT);
-    }
-
-}
