--- src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -1,311 +1,0 @@
-package seedu.address.logic.commands;
-
-import static java.util.Objects.requireNonNull;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.testutil.Assert.assertThrows;
-
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.function.Predicate;
-
-import org.junit.jupiter.api.Test;
-
-import javafx.collections.ObservableList;
-import seedu.address.commons.core.GuiSettings;
-import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
-import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyBuyerAddressBook;
-import seedu.address.model.ReadOnlySellerAddressBook;
-import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.buyer.Buyer;
-import seedu.address.model.client.Client;
-import seedu.address.model.seller.Seller;
-import seedu.address.testutil.ClientBuilder;
-
-public class AddCommandTest {
-
-    @Test
-    public void constructor_nullclient_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> new AddCommand(null));
-    }
-
-    @Test
-    public void execute_clientAcceptedByModel_addSuccessful() throws Exception {
-        ModelStubAcceptingClientAdded modelStub = new ModelStubAcceptingClientAdded();
-        Client validClient = new ClientBuilder().build();
-
-        CommandResult commandResult = new AddCommand(validClient).execute(modelStub);
-
-        assertEquals(String.format(AddCommand.MESSAGE_SUCCESS, validClient), commandResult.getFeedbackToUser());
-        assertEquals(Arrays.asList(validClient), modelStub.clientsAdded);
-    }
-
-    @Test
-    public void execute_duplicateclient_throwsCommandException() {
-        Client validClient = new ClientBuilder().build();
-        AddCommand addCommand = new AddCommand(validClient);
-        ModelStub modelStub = new ModelStubWithClient(validClient);
-
-        assertThrows(CommandException.class, AddCommand.MESSAGE_DUPLICATE_CLIENT, () -> addCommand.execute(modelStub));
-    }
-
-    @Test
-    public void equals() {
-        Client alice = new ClientBuilder().withName("Alice").build();
-        Client bob = new ClientBuilder().withName("Bob").build();
-        AddCommand addAliceCommand = new AddCommand(alice);
-        AddCommand addBobCommand = new AddCommand(bob);
-
-        // same object -> returns true
-        assertTrue(addAliceCommand.equals(addAliceCommand));
-
-        // same values -> returns true
-        AddCommand addAliceCommandCopy = new AddCommand(alice);
-        assertTrue(addAliceCommand.equals(addAliceCommandCopy));
-
-        // different types -> returns false
-        assertFalse(addAliceCommand.equals(1));
-
-        // null -> returns false
-        assertFalse(addAliceCommand.equals(null));
-
-        // different client -> returns false
-        assertFalse(addAliceCommand.equals(addBobCommand));
-    }
-
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addClient(Client client) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasClient(Client client) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasBuyer(Buyer buyer) {
-            throw new AssertionError("This method should not be called");
-        }
-
-        @Override
-        public boolean hasSeller(Seller seller) {
-            throw new AssertionError("This method should not be called");
-        }
-
-        @Override
-        public void deleteClient(Client target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setClient(Client target, Client editedClient) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void sortFilteredBuyerList(String comparedItem, String order) {
-            throw new AssertionError("This method should not be called");
-        }
-
-
-
-        @Override
-        public ObservableList<Client> getFilteredClientList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredClientList(Predicate<Client> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addBuyer(Buyer buyer) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addSeller(Seller seller) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getSellerAddressBookFilePath() {
-            return null;
-        }
-
-        @Override
-        public void setSellerAddressBookFilePath(Path sellerAddressBookFilePath) {
-
-        }
-
-        @Override
-        public void setSellerAddressBook(ReadOnlySellerAddressBook sellerAddressBook) {
-
-        }
-
-        @Override
-        public ReadOnlySellerAddressBook getSellerAddressBook() {
-            return null;
-        }
-
-        @Override
-        public void deleteSeller(Seller target) {
-
-        }
-
-        @Override
-        public void setSeller(Seller target, Seller editedSeller) {
-
-        }
-
-        @Override
-        public ObservableList<Seller> getFilteredSellerList() {
-            return null;
-        }
-
-        @Override
-        public void updateFilteredSellerList(Predicate<Seller> predicate) {
-
-        }
-
-        @Override
-        public void sortFilteredSellerList(String comparator, String order) {
-            throw new AssertionError("This method should not be called");
-        }
-
-        // buyer //
-        @Override
-        public Path getBuyerAddressBookFilePath() {
-            return null;
-        }
-
-        @Override
-        public void setBuyerAddressBookFilePath(Path buyerAddressBookFilePath) {
-
-        }
-
-        @Override
-        public void setBuyerAddressBook(ReadOnlyBuyerAddressBook buyerAddressBook) {
-
-        }
-
-        @Override
-        public ReadOnlyBuyerAddressBook getBuyerAddressBook() {
-            return null;
-        }
-
-        @Override
-        public void deleteBuyer(Buyer target) {
-
-        }
-
-        @Override
-        public void setBuyer(Buyer target, Buyer editedBuyer) {
-
-        }
-
-        @Override
-        public ObservableList<Buyer> getFilteredBuyerList() {
-            return null;
-        }
-
-        @Override
-        public void updateFilteredBuyerList(Predicate<Buyer> predicate) {
-
-        }
-    }
-
-    /**
-     * A Model stub that contains a single client.
-     */
-    private class ModelStubWithClient extends ModelStub {
-        private final Client client;
-
-        ModelStubWithClient(Client client) {
-            requireNonNull(client);
-            this.client = client;
-        }
-
-        @Override
-        public boolean hasClient(Client client) {
-            requireNonNull(client);
-            return this.client.isSameclient(client);
-        }
-    }
-
-    /**
-     * A Model stub that always accept the client being added.
-     */
-    private class ModelStubAcceptingClientAdded extends ModelStub {
-        final ArrayList<Client> clientsAdded = new ArrayList<>();
-
-        @Override
-        public boolean hasClient(Client client) {
-            requireNonNull(client);
-            return clientsAdded.stream().anyMatch(client::isSameclient);
-        }
-
-        @Override
-        public void addClient(Client client) {
-            requireNonNull(client);
-            clientsAdded.add(client);
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            return new AddressBook();
-        }
-    }
-
-}
