+++ src/test/java/seedu/address/logic/commands/FindCommandTest.java
@@ -1,87 +1,87 @@
-package seedu.address.logic.commands;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.commons.core.Messages.MESSAGE_CLIENTS_LISTED_OVERVIEW;
-import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.testutil.TypicalClients.CARL;
-import static seedu.address.testutil.TypicalClients.ELLE;
-import static seedu.address.testutil.TypicalClients.FIONA;
-import static seedu.address.testutil.TypicalClients.getTypicalAddressBook;
-
-import java.util.Arrays;
-import java.util.Collections;
-
-import org.junit.jupiter.api.Test;
-
-import seedu.address.model.BuyerAddressBook;
-import seedu.address.model.Model;
-import seedu.address.model.ModelManager;
-import seedu.address.model.SellerAddressBook;
-import seedu.address.model.UserPrefs;
-import seedu.address.model.client.NameContainsKeywordsPredicate;
-
-/**
- * Contains integration tests (interaction with the Model) for {@code FindCommand}.
- */
-public class FindCommandTest {
-    private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs(), new SellerAddressBook(),
-            new BuyerAddressBook());
-    private Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs(),
-            new SellerAddressBook(), new BuyerAddressBook());
-
-    @Test
-    public void equals() {
-        NameContainsKeywordsPredicate firstPredicate =
-                new NameContainsKeywordsPredicate(Collections.singletonList("first"));
-        NameContainsKeywordsPredicate secondPredicate =
-                new NameContainsKeywordsPredicate(Collections.singletonList("second"));
-
-        FindCommand findFirstCommand = new FindCommand(firstPredicate);
-        FindCommand findSecondCommand = new FindCommand(secondPredicate);
-
-        // same object -> returns true
-        assertTrue(findFirstCommand.equals(findFirstCommand));
-
-        // same values -> returns true
-        FindCommand findFirstCommandCopy = new FindCommand(firstPredicate);
-        assertTrue(findFirstCommand.equals(findFirstCommandCopy));
-
-        // different types -> returns false
-        assertFalse(findFirstCommand.equals(1));
-
-        // null -> returns false
-        assertFalse(findFirstCommand.equals(null));
-
-        // different client -> returns false
-        assertFalse(findFirstCommand.equals(findSecondCommand));
-    }
-
-    @Test
-    public void execute_zeroKeywords_noclientFound() {
-        String expectedMessage = String.format(MESSAGE_CLIENTS_LISTED_OVERVIEW, 0);
-        NameContainsKeywordsPredicate predicate = preparePredicate(" ");
-        FindCommand command = new FindCommand(predicate);
-        expectedModel.updateFilteredClientList(predicate);
-        assertCommandSuccess(command, model, expectedMessage, expectedModel);
-        assertEquals(Collections.emptyList(), model.getFilteredClientList());
-    }
-
-    @Test
-    public void execute_multipleKeywords_multipleclientsFound() {
-        String expectedMessage = String.format(MESSAGE_CLIENTS_LISTED_OVERVIEW, 3);
-        NameContainsKeywordsPredicate predicate = preparePredicate("Kurz Elle Kunz");
-        FindCommand command = new FindCommand(predicate);
-        expectedModel.updateFilteredClientList(predicate);
-        assertCommandSuccess(command, model, expectedMessage, expectedModel);
-        assertEquals(Arrays.asList(CARL, ELLE, FIONA), model.getFilteredClientList());
-    }
-
-    /**
-     * Parses {@code userInput} into a {@code NameContainsKeywordsPredicate}.
-     */
-    private NameContainsKeywordsPredicate preparePredicate(String userInput) {
-        return new NameContainsKeywordsPredicate(Arrays.asList(userInput.split("\\s+")));
-    }
-}
+//package seedu.address.logic.commands;
+//
+//import static org.junit.jupiter.api.Assertions.assertEquals;
+//import static org.junit.jupiter.api.Assertions.assertFalse;
+//import static org.junit.jupiter.api.Assertions.assertTrue;
+//import static seedu.address.commons.core.Messages.MESSAGE_CLIENTS_LISTED_OVERVIEW;
+//import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
+//import static seedu.address.testutil.TypicalClients.CARL;
+//import static seedu.address.testutil.TypicalClients.ELLE;
+//import static seedu.address.testutil.TypicalClients.FIONA;
+//import static seedu.address.testutil.TypicalClients.getTypicalAddressBook;
+//
+//import java.util.Arrays;
+//import java.util.Collections;
+//
+//import org.junit.jupiter.api.Test;
+//
+//import seedu.address.model.BuyerAddressBook;
+//import seedu.address.model.Model;
+//import seedu.address.model.ModelManager;
+//import seedu.address.model.SellerAddressBook;
+//import seedu.address.model.UserPrefs;
+//import seedu.address.model.client.NameContainsKeywordsPredicate;
+//
+///**
+// * Contains integration tests (interaction with the Model) for {@code FindCommand}.
+// */
+//public class FindCommandTest {
+//    private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs(), new SellerAddressBook(),
+//            new BuyerAddressBook());
+//    private Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs(),
+//            new SellerAddressBook(), new BuyerAddressBook());
+//
+//    @Test
+//    public void equals() {
+//        NameContainsKeywordsPredicate firstPredicate =
+//                new NameContainsKeywordsPredicate(Collections.singletonList("first"));
+//        NameContainsKeywordsPredicate secondPredicate =
+//                new NameContainsKeywordsPredicate(Collections.singletonList("second"));
+//
+//        FindCommand findFirstCommand = new FindCommand(firstPredicate);
+//        FindCommand findSecondCommand = new FindCommand(secondPredicate);
+//
+//        // same object -> returns true
+//        assertTrue(findFirstCommand.equals(findFirstCommand));
+//
+//        // same values -> returns true
+//        FindCommand findFirstCommandCopy = new FindCommand(firstPredicate);
+//        assertTrue(findFirstCommand.equals(findFirstCommandCopy));
+//
+//        // different types -> returns false
+//        assertFalse(findFirstCommand.equals(1));
+//
+//        // null -> returns false
+//        assertFalse(findFirstCommand.equals(null));
+//
+//        // different client -> returns false
+//        assertFalse(findFirstCommand.equals(findSecondCommand));
+//    }
+//
+//    @Test
+//    public void execute_zeroKeywords_noclientFound() {
+//        String expectedMessage = String.format(MESSAGE_CLIENTS_LISTED_OVERVIEW, 0);
+//        NameContainsKeywordsPredicate predicate = preparePredicate(" ");
+//        FindCommand command = new FindCommand(predicate);
+//        expectedModel.updateFilteredClientList(predicate);
+//        assertCommandSuccess(command, model, expectedMessage, expectedModel);
+//        assertEquals(Collections.emptyList(), model.getFilteredClientList());
+//    }
+//
+//    @Test
+//    public void execute_multipleKeywords_multipleclientsFound() {
+//        String expectedMessage = String.format(MESSAGE_CLIENTS_LISTED_OVERVIEW, 3);
+//        NameContainsKeywordsPredicate predicate = preparePredicate("Kurz Elle Kunz");
+//        FindCommand command = new FindCommand(predicate);
+//        expectedModel.updateFilteredClientList(predicate);
+//        assertCommandSuccess(command, model, expectedMessage, expectedModel);
+//        assertEquals(Arrays.asList(CARL, ELLE, FIONA), model.getFilteredClientList());
+//    }
+//
+//    /**
+//     * Parses {@code userInput} into a {@code NameContainsKeywordsPredicate}.
+//     */
+//    private NameContainsKeywordsPredicate preparePredicate(String userInput) {
+//        return new NameContainsKeywordsPredicate(Arrays.asList(userInput.split("\\s+")));
+//    }
+//}
