+++ src/test/java/seedu/address/logic/commands/ListCommandTest.java
@@ -1,43 +1,43 @@
-package seedu.address.logic.commands;
-
-import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.logic.commands.CommandTestUtil.showClientAtIndex;
-import static seedu.address.testutil.TypicalClients.getTypicalAddressBook;
-import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_CLIENT;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import seedu.address.model.BuyerAddressBook;
-import seedu.address.model.Model;
-import seedu.address.model.ModelManager;
-import seedu.address.model.SellerAddressBook;
-import seedu.address.model.UserPrefs;
-
-/**
- * Contains integration tests (interaction with the Model) and unit tests for ListCommand.
- */
-public class ListCommandTest {
-
-    private Model model;
-    private Model expectedModel;
-
-    @BeforeEach
-    public void setUp() {
-        model = new ModelManager(getTypicalAddressBook(), new UserPrefs(), new SellerAddressBook(),
-                new BuyerAddressBook());
-        expectedModel = new ModelManager(model.getAddressBook(), new UserPrefs(), new SellerAddressBook(),
-                new BuyerAddressBook());
-    }
-
-    @Test
-    public void execute_listIsNotFiltered_showsSameList() {
-        assertCommandSuccess(new ListCommand(), model, ListCommand.MESSAGE_SUCCESS, expectedModel);
-    }
-
-    @Test
-    public void execute_listIsFiltered_showsEverything() {
-        showClientAtIndex(model, INDEX_FIRST_CLIENT);
-        assertCommandSuccess(new ListCommand(), model, ListCommand.MESSAGE_SUCCESS, expectedModel);
-    }
-}
+//package seedu.address.logic.commands;
+//
+//import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
+//import static seedu.address.logic.commands.CommandTestUtil.showClientAtIndex;
+//import static seedu.address.testutil.TypicalClients.getTypicalAddressBook;
+//import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_CLIENT;
+//
+//import org.junit.jupiter.api.BeforeEach;
+//import org.junit.jupiter.api.Test;
+//
+//import seedu.address.model.BuyerAddressBook;
+//import seedu.address.model.Model;
+//import seedu.address.model.ModelManager;
+//import seedu.address.model.SellerAddressBook;
+//import seedu.address.model.UserPrefs;
+//
+///**
+// * Contains integration tests (interaction with the Model) and unit tests for ListCommand.
+// */
+//public class ListCommandTest {
+//
+//    private Model model;
+//    private Model expectedModel;
+//
+//    @BeforeEach
+//    public void setUp() {
+//        model = new ModelManager(getTypicalAddressBook(), new UserPrefs(), new SellerAddressBook(),
+//                new BuyerAddressBook());
+//        expectedModel = new ModelManager(model.getAddressBook(), new UserPrefs(), new SellerAddressBook(),
+//                new BuyerAddressBook());
+//    }
+//
+//    @Test
+//    public void execute_listIsNotFiltered_showsSameList() {
+//        assertCommandSuccess(new ListCommand(), model, ListCommand.MESSAGE_SUCCESS, expectedModel);
+//    }
+//
+//    @Test
+//    public void execute_listIsFiltered_showsEverything() {
+//        showClientAtIndex(model, INDEX_FIRST_CLIENT);
+//        assertCommandSuccess(new ListCommand(), model, ListCommand.MESSAGE_SUCCESS, expectedModel);
+//    }
+//}
