+++ src/test/java/seedu/address/logic/parser/AddBuyerCommandParserTest.java
@@ -1,21 +1,21 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_NAME_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_PHONE_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_TAG_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.PHONE_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.PHONE_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.PREAMBLE_NON_EMPTY;
-import static seedu.address.logic.commands.CommandTestUtil.PREAMBLE_WHITESPACE;
-import static seedu.address.logic.commands.CommandTestUtil.TAG_DESC_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.TAG_DESC_HUSBAND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.INVALID_NAME_DESC;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.INVALID_PHONE_DESC;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.INVALID_TAG_DESC;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.NAME_DESC_AMY;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.NAME_DESC_BOB;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.PHONE_DESC_AMY;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.PHONE_DESC_BOB;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.PREAMBLE_NON_EMPTY;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.PREAMBLE_WHITESPACE;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.TAG_DESC_FRIEND;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.TAG_DESC_HUSBAND;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.VALID_NAME_BOB;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.VALID_PHONE_BOB;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.VALID_TAG_FRIEND;
+import static seedu.address.logic.commands.BuyerCommandTestUtil.VALID_TAG_HUSBAND;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
 import static seedu.address.testutil.TypicalBuyers.AMY;
