+++ src/test/java/seedu/address/model/seller/SellerTest.java
@@ -5,9 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.logic.commands.SellerCommandTestUtil.VALID_NAME_BOB;
+import static seedu.address.logic.commands.SellerCommandTestUtil.VALID_PHONE_BOB;
+import static seedu.address.logic.commands.SellerCommandTestUtil.VALID_TAG_HUSBAND;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalSellers.BOB;
 import static seedu.address.testutil.TypicalSellers.YUQI;
